requirements.txt
PyQt5==5.15.9
PyYAML==6.0
requests==2.28.2
swagger-parser==1.0.3
jsonschema==4.17.3
python-dateutil==2.8.2
urllib3==1.26.15
pytest==7.3.1
pytest-html==3.2.0
faker==18.7.0
